Checking totp also, moving totp to api common.
