+++ crates/api/Cargo.toml
@@ -36,7 +36,6 @@ url = { workspace = true }
 regex = { workspace = true }
 hound = "3.5.1"
 sitemap-rs = "0.2.2"
-totp-rs = { version = "5.6.0", features = ["gen_secret", "otpauth"] }
 diesel-async = { workspace = true, features = ["deadpool", "postgres"] }
 
 [dev-dependencies]
