+++ crates/api/src/lib.rs
@@ -1,13 +1,11 @@
 use base64::{engine::general_purpose::STANDARD_NO_PAD as base64, Engine};
 use captcha::Captcha;
-use lemmy_db_views::structs::LocalUserView;
 use lemmy_utils::{
   error::{LemmyErrorExt, LemmyErrorType, LemmyResult},
   utils::slurs::check_slurs,
 };
 use regex::Regex;
 use std::io::Cursor;
-use totp_rs::{Secret, TOTP};
 
 pub mod comment;
 pub mod community;
@@ -70,62 +68,3 @@ pub(crate) fn check_report_reason(reason: &str, slur_regex: &Regex) -> LemmyResu
   }
 }
 
-pub(crate) fn check_totp_2fa_valid(
-  local_user_view: &LocalUserView,
-  totp_token: &Option<String>,
-  site_name: &str,
-) -> LemmyResult<()> {
-  // Throw an error if their token is missing
-  let token = totp_token
-    .as_deref()
-    .ok_or(LemmyErrorType::MissingTotpToken)?;
-  let secret = local_user_view
-    .local_user
-    .totp_2fa_secret
-    .as_deref()
-    .ok_or(LemmyErrorType::MissingTotpSecret)?;
-
-  let totp = build_totp_2fa(site_name, &local_user_view.person.name, secret)?;
-
-  let check_passed = totp.check_current(token)?;
-  if !check_passed {
-    return Err(LemmyErrorType::IncorrectTotpToken.into());
-  }
-
-  Ok(())
-}
-
-pub(crate) fn generate_totp_2fa_secret() -> String {
-  Secret::generate_secret().to_string()
-}
-
-fn build_totp_2fa(hostname: &str, username: &str, secret: &str) -> LemmyResult<TOTP> {
-  let sec = Secret::Raw(secret.as_bytes().to_vec());
-  let sec_bytes = sec
-    .to_bytes()
-    .with_lemmy_type(LemmyErrorType::CouldntParseTotpSecret)?;
-
-  TOTP::new(
-    totp_rs::Algorithm::SHA1,
-    6,
-    1,
-    30,
-    sec_bytes,
-    Some(hostname.to_string()),
-    username.to_string(),
-  )
-  .with_lemmy_type(LemmyErrorType::CouldntGenerateTotp)
-}
-
-#[cfg(test)]
-mod tests {
-
-  use super::*;
-
-  #[test]
-  fn test_build_totp() {
-    let generated_secret = generate_totp_2fa_secret();
-    let totp = build_totp_2fa("lemmy.ml", "my_name", &generated_secret);
-    assert!(totp.is_ok());
-  }
-}
