+++ crates/api/src/local_user/generate_totp_secret.rs
@@ -1,7 +1,10 @@
-use crate::{build_totp_2fa, generate_totp_2fa_secret};
 use activitypub_federation::config::Data;
 use actix_web::web::Json;
-use lemmy_api_common::{context::LemmyContext, person::GenerateTotpSecretResponse};
+use lemmy_api_common::{
+  context::LemmyContext,
+  person::GenerateTotpSecretResponse,
+  utils::{build_totp_2fa, generate_totp_2fa_secret},
+};
 use lemmy_db_schema::source::{
   local_user::{LocalUser, LocalUserUpdateForm},
   site::Site,
