+++ crates/api_common/Cargo.toml
@@ -40,6 +40,7 @@ full = [
   "extism-convert",
   "once_cell",
   "serde_json",
+  "totp-rs",
   "lemmy_db_schema_file/full",
 ]
 
@@ -80,6 +81,10 @@ webmention = { version = "0.6.0", optional = true }
 extism = { git = "https://github.com/extism/extism.git", branch = "pool", optional = true }
 extism-convert = { git = "https://github.com/extism/extism.git", branch = "pool", optional = true }
 once_cell = { version = "1.21.0", optional = true }
+totp-rs = { version = "5.6.0", optional = true, features = [
+  "gen_secret",
+  "otpauth",
+] }
 
 [dev-dependencies]
 serial_test = { workspace = true }
